#67 automcomplete device number, cleanup

+++ backend/src/main/java/at/camconnect/dtos/DeviceDTO.java
@@ -2,5 +2,5 @@ package at.camconnect.dtos;
 
 import at.camconnect.model.DeviceType;
 
-public record DeviceDTO(String serial, String number, String note, int type_id) {
+public record DeviceDTO(String serial, String number, String note, Long type_id) {
 }
